ARCHIVE_20250606/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.2
aiosignal==1.3.1
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anthropic-cli==0.2.3
anyio==4.8.0
appnope==0.1.4
argcomplete==1.10.3
asttokens==2.4.1
asyncio==3.4.3
attrs==24.2.0
Authlib==1.5.2
backcall==0.2.0
bandit==1.8.3
bcrypt==4.3.0
beautifulsoup4==4.13.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
boolean.py==5.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
compressed-rtf==1.0.7
contourpy==1.3.1
coverage==7.6.7
cryptography==44.0.2
cycler==0.12.1
cyclonedx-python-lib==9.1.0
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
docx2txt==0.9
dparse==0.6.4
ebcdic==1.1.1
email_validator==2.2.0
eval_type_backport==0.2.0
executing==2.1.0
extract-msg==0.28.7
faiss-cpu==1.10.0
fastapi==0.115.8
fastavro==1.10.0
fastjsonschema==2.20.0
filelock==3.16.1
flake8==7.2.0
flake8-bandit==4.1.1
flake8-bugbear==24.12.12
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.70.0
grpcio-status==1.63.0rc1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.2
identify==2.6.9
idna==3.10
IMAPClient==2.1.0
iniconfig==2.0.0
ipython==8.12.3
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
langchain==0.3.22
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.22
license-expression==30.4.1
llama_cpp_python==0.3.8
llvmlite==0.44.0
lxml==5.3.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistralai==1.7.0
mistune==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
narwhals==1.37.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
olefile==0.47
openai==1.70.0
orjson==3.10.16
packageurl-python==0.16.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.1
pdf2image==1.17.0
pdfminer.six==20191110
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.2.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.26.0
protobuf==4.25.6
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-serializable==2.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.27.1
pydeck==0.9.1
pyflakes==3.3.2
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.1
pypdf==5.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-multipart==0.0.20
python-pptx==0.6.23
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
safety==3.3.1
safety-schemas==0.0.11
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
sentry-sdk==2.25.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.40
sse-starlette==2.1.0
stack-data==0.6.3
starlette==0.45.3
stevedore==5.4.1
streamlit==1.45.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
xlrd==1.2.0
XlsxWriter==3.2.2
yarg==0.1.9
yarl==1.17.1
zstandard==0.23.0
